Fix thread name length in exit tests' call to `pthread_setname_np()`. (#367)
On Linux, the maximum length of a thread's name is defined by
`TASK_COMM_LEN` to be 16 bytes (including the trailing null character.)
We currently manually create one thread on Linux in order to monitor for
termination of exit tests' child processes. This PR changes the name of
the monitor thread on Linux to one that is less than 16 characters long.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.