Fix a Linux bug where an error during `waitid()` could cause us to miss a continuation. (#369)
The logic for exit test handling currently calls `waitid()` but doesn't
loop on error (`EINTR` or `ECHILD`) so it could cause a missed
continuation in corner cases. This PR refactors (_again_) the relevant
code so that it loops and is shared with the
Darwin/dispatch-source-based implementation.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.