Remove `Bug.Relationship`.
This PR removes the `Bug.Relationship` type and corresponding `relationship`
property and accoutrements.

We had planned to expose this information so that developers could provide
finer-grained metadata about related bugs that could be surfaced in `swift test`
or IDEs like VSCode.

In practice, we've struggled to figure out how this information could actually
be used: a tool like `swift test` doesn't have write access to a developer's
bug tracking software, so it couldn't go update a bug with a comment saying
"this issue has started occurring again" or something like that.

We are interested in re-exposing this API or something similar in the future,
but for our initial release we are removing it.