Rename `Expression` to `__Expression`.
This PR renames the `Expression` type, which represents source code and syntax
captured during macro expansion, so that it does not conflict with Foundation's
[`Expression`](https://github.com/apple/swift-foundation/blob/main/Sources/FoundationEssentials/Predicate/Expression.swift)
type used with predicates.

We also expose a typealias to `__Expression` for use when integrating with tools
that care about captured source code (e.g. to present it in their output.)

Resolves rdar://125988077.