Stop using the `SWIFT_PM_SUPPORTS_SWIFT_TESTING` flag. (#376)
When we were first bringing up experimental support for swift-testing in
Swift Package Manager, I added the flag
`SWIFT_PM_SUPPORTS_SWIFT_TESTING` to the build process for test targets
so that we could detect if there was built-in support or not. Now that
we've announced we're aligning our release with Swift 6, we can instead
just check if the compiler is Swift 6 or later. Very early Swift 5.11
toolchain builds did not have support, but they have long since been
superseded.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.