Provide an ABI-friendly entry point function that takes/produces JSON. (#360)
This PR introduces an **experimental** function, intended to eventually
become part of our stable ABI, named `swt_copyABIEntryPoint_v0()`. This
function can be looked up at runtime and used to hook into the testing
library when a test target is loaded into the current process. It takes
a JSON-encoded structure as input and yields JSON-encoded events/event
contexts via a callback.

The input is encoded from a type we've tentatively named
`__CommandLineArguments_v0`. This type conforms to `Codable` and
represents the parsed command-line arguments that the package can
understand. Normally, swift-testing will construct an instance of this
type from the command-line arguments to the current process, but a
caller can pass in a different instance to the ABI entry point instead.

The output consists of `Event.Snapshot` and `Event.Context.Snapshot`
values encoded in a single JSON object. The format of this JSON is
identical to what is emitted when `--experimental-event-stream-output`
is specified at the command line.

`__swiftPMEntryPoint()` has also been modified to use the same
`__CommandLineArguments_v0` type as input and as much of the same
implementation as possible.

Because trying to load a Swift function dynamically at runtime is Hard™,
`swt_copyABIEntryPoint_v0()` is a C function that yields the "real"
Swift function. This indirection also gives us the flexibility to
produce a thick Swift closure rather than a thin function.

> [!WARNING]
> The signature of structure of `swt_copyABIEntryPoint_v0`, the
structure of `__CommandLineArguments_v0`, how it encodes to/from JSON,
and all other details about it are still subject to change. The hope is
that it becomes a stable ABI surface _eventually_, but we're not there
yet. :)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.