Adopt swift-syntax-600. (#338)
This PR changes our swift-syntax dependency from swift-syntax-510 to
swift-syntax-600.

We rely on several upcoming Swift 6 features, including swift-syntax's
`lexicalContext` property, in order to correctly build and run tests.
This PR migrates us to the latest prerelease swift-syntax-600 package so
that we can take advantage of said feature.

swift-testing clients that have an explicit or inherited dependency on
swift-syntax-510 or earlier may fail to build as a result of this
change. As a reminder, swift-testing is an experimental package and we
do not guarantee compatibility with older versions of the toolchain or
related packages.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.