Re-enable some pre-swift-syntax-600 code.
The Swift 5.10 compiler does not populate the new `lexicalContext` property during macro expansion. This PR ensures we maintain minimal 5.10 compatibility by continuing to use our whitespace trick to "get" the containing type of a test or suite.

This change can be reverted once support for the Swift 5.10 compiler is dropped.