Refactor type enumeration loop.
This PR reworks `swt_enumerateTypes()` and adds a Swift wrapper so that less
boilerplate is needed at call sites. Since we're always using it to look for
types with a particular substring in their names, I've dropped the name-matcher
callback and replaced it with a simple "give me a substring to look for"-type
argument.