Return the ABI entry point from `swt_copyABIEntryPoint_v0()` normally.
This PR changes the calling convention of `swt_copyABIEntryPoint_v0()` to return
a pointer to the ABI entry point rather than passing it back through an
out-parameter. The previous signature was an attempt to allow using
`withUnsafeTemporaryAllocation()` (to avoid a heap allocation) but that's a
premature optimization and makes the function harder to use.