Return the ABI entry point from `swt_copyABIEntryPoint_v0()` normally. (#389)
This PR changes the calling convention of `swt_copyABIEntryPoint_v0()`
to return a pointer to the ABI entry point rather than passing it back
through an out-parameter. The previous signature was an attempt to allow
using `withUnsafeTemporaryAllocation()` (to avoid a heap allocation) but
that's a premature optimization and makes the function harder to use.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.