Add an experimental CLI argument for passing a configuration value as JSON.
This PR adds a new CLI argument, `--experimental-configuration-path`, which can
be used to specify the path to a file containing an instance of
`__CommandLineArguments_v0` encoded as JSON. `/dev/stdin` or a named pipe can be
specified if desired.

Changes to SwiftPM will be needed to support passing this argument all the way
through from `swift test`.