Rename `TestingInternals` to `_TestingInternals`. (#404)
This module is not meant to be accessible outside of swift-testing,
however it may be reachable by clients inadvertently. To minimize the
likelihood of that, rename the module so it doesn't show up in
documentation or Xcode autocomplete contexts.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.