Define and implement a strawman stable JSON schema for output. (#383)
This PR defines a JSON schema for events output from swift-testing via
either `--experimental-event-stream-output` or `abiEntryPoint_v0()`. The
JSON schema needs to be formally reviewed separately.

We are already writing an experimental JSON stream using "snapshot"
types and we don't want to break folks experimenting with it before we
have a final JSON schema, so for now that remains the default. To opt
into the new schema, pass `--experimental-event-stream-version 0` (note
that this argument is not forwarded from `swift test`. See
https://github.com/apple/swift-package-manager/pull/7534.)

Resolves #368.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.