Add temporary shims to `Bug`'s `Codable` implementation to support proposed new layout.
We're proposing a new set of stored properties for `Bug` in #412. This new
layout is incompatible with the old one when using `Codable`. This change adds a
(very!) temporary workaround to ensure that bugs encoded with the current layout
can be decoded with the new one once it is accepted and merged in.