Add temporary shims to `Bug`'s `Codable` implementation to support proposed new layout. (#419)
We're proposing a new set of stored properties for `Bug` in #412. This
new layout is incompatible with the old one when using `Codable`. This
change adds a (very!) temporary workaround to ensure that bugs encoded
with the current layout can be decoded with the new one once it is
accepted and merged in.

This implementation can be removed when #412 is merged.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.