Remove string literal tags. (#420)
As discussed [in the forums](https://forums.swift.org/t/removal-of-string-literal-tag-definitions/71667),
we are removing string literal tags. This PR removes the interfaces for
them. It preserves `Tag.Kind` because that may prove to be a useful
abstraction in the future, e.g. if we come up with a new kind of tag
that isn't a member of `Tag`.

Resolves rdar://125730118.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.