Move process spawning to its own file so it can be reused in harness logic.
This PR speculatively moves the process-spawning logic used by the experimental
exit tests feature to its own file so that we can reuse it if/when we build out
a more complex test harness feature.

We aren't currently planning to build out such a feature, but we suspect we'll
need something in this area, and we'd prefer to reuse this code if possible
rather than write equivalent code all over again.