Ensure `@Test(arguments:)` macro expansion trims trivia from argument expressions (#448)
This fixes a build error which can arise if the `@Test(arguments:)`
attribute applied to a parameterized test function includes trailing
trivia such as a `//`-style line comment on the end of the line. For
example:

```swift
@Test(
  arguments: [...] // some comment
)
func example(...) { ... }
```

This regressed with my changes in #366 which began surrounding the
argument expressions in a closure expression, in order to facilitate
lazy evaluation. The fix ensures that the wrapped expression is trimmed,
in addition to the outer closure (which was, and still is, trimmed).

### Motivation:

Comments in this position may not be used often, but can be useful to
provide context about the role or meaning of a particular collection of
arguments passed to a parameterized test. So we'd like to continue
supporting this.

### Modifications:

- Trim trivia from the argument expression included in the `@Test` macro
expansion code.
- Add test reproducing and validating this scenario.

### Result:

Parameterized test functions with trailing line comments on the
`arguments:` line now produce valid macro expansion code and build
successfully.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

Resolves rdar://128609944