Avoid infinite recursion when snapshotting objects which may have cyclic references (#449)
This fixes an issue which can cause a crash due to infinite recursion
when taking snapshots of values which have cyclic or recursive object
references.

### Modifications:

- Begin keeping track of the objects seen while recursively snapshotting
child values, and stop recursion if we encounter a previously-seen
object. The logic is carefully constructed to limit the overhead
involved in managing this data and limiting it to only apply to object
types. See the code comment for details.
- Added a new unit test suite validating this fix, as well as the
general behavior.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

Resolves rdar://124726503