Move dynamic loader abstraction up to Swift. (#452)
This PR moves the library's calls to `dlsym()` and `GetProcAddress()` up
to Swift. Previously, they were implemented in C++, but as a general
policy we want as much of swift-testing to be implemented in Swift as
possible.

The original reason for implementing it in C++ was its dependence on
`RTLD_DEFAULT` which, on Linux, is not directly available to Swift.
However, the value of this constant is defined as `0` by glibc and we
can simply redefine it in Swift for the purposes of this function call.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.