Throw errors from the ABI entry point instead of crashing.
This PR changes the signature of the ABI entry point function to be throwing.
Any existing early adopters will need to add `try` to their calls to it (not to
the function that yields the Swift entry point, but to the yielded Swift entry
point.)

Callers that are uninterested in errors that occur during setup can use `try!`
or `try?` as appropriate.