Warn about the use of `.serialized` with non-parameterized test functions (#441)
This PR adds a compile-time diagnostic if `.serialized` is used with
`@Test` for a test function that is not parameterized. Such uses are
nonsensical, because a non-parameterized test function has nothing to
serialize against. The trait operates "inward", meaning that applying it
to a _suite_ affects the tests in that suite, and applying it to a
_parameterized test function_ applies to the test cases of that
function.

For example:

```swift
@Test(.serialized) // ⚠️ Warning: Trait '.serialized' has no effect when used
                   // with a non-parameterized test function
func f() { /* ... */ }
```

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.

---------

Co-authored-by: Stuart Montgomery <smontgomery@apple.com>