Don't emit JSON blobs for test cases if a test function isn't parameterized. (#482)
Internally, Swift Testing models non-parameterized test functions as
parameterized over a single input, `()`. This is a very useful
abstraction internally, but when it comes to the JSON output for
parameterized tests, it can be confusing or surprising to see
information about a test case that was synthesized by the library.

This PR suppresses events and metadata for the single test case of a
non-parameterized test function.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.