Add experimental overload of `confirmation()` that takes a range.
Wouldn't it be useful to be able to specify a range of expected counts for a confirmation rather than a specific number? For instance, if you expect an event to occur exactly zero or one times:

```swift
await confirmation(expectedCount: 0 ... 1) { thingHappened in
  // ...
}
```

Or maybe you expect that an event will occur _at least_ 10 times, but maybe many more?

```swift
await confirmation(expectedCount: 10...) { thingHappened in
  // ...
}
```

Well... good news, everyone! Here's an experimental overload of `confirmation()` that allows you to do exactly that!