Support building macro plugin as a library instead of executable via CMake (#517)
This adds support in the CMake rules for building the `TestingMacros`
macro plugin target as a library instead of an executable.

### Motivation:

In preparation for adding Swift Testing to Swift.org toolchains (as
[mentioned](https://github.com/swiftlang/swift-evolution/blob/main/visions/swift-testing.md#distribution)
in the vision document), this modifies the project's CMake rules to
allow building its macro plugin in a way that is consistent with others
in the toolchain — as a shared library rather than an executable. Macros
in the toolchain can be libraries which are directly loaded by the
compiler and avoid the IPC overhead of being a separate executable.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.