Avoid fork-bombing ourselves when hosted by `swift-test`.
This PR amends exit tests' child process spawning logic to ensure that they
don't trigger a fork bomb or fail outright when the test process is hosted by
Swift Package Manager (as in [this branch](https://github.com/swiftlang/swift-package-manager/tree/jgrynspan/one-test-product-two-libraries)).

Exit tests are an experimental feature, but one that we use internally in our
own test suite, so we need to make sure our tests continue to function if/when
that SwiftPM branch is merged.