Change the default JSON schema version to v0.
This PR changes the default JSON schema version for event stream output to
version 0 (the version documented in #479.) Previously, if a caller did not
specify a version number for the event stream, it would default to encoding
"snapshots" of various values from the testing library.

Xcode 16 Beta 1 uses this legacy format, so to continue supporting newer betas
of Xcode 16 transitionally, this PR temporarily enables a "-1" JSON schema
version that continues to emit snapshot encodings. This JSON schema version will
be removed (along with the snapshots code in general) in a future update once
Xcode has migrated.

Resolves #527.