[SWT-0002] A stable JSON-based ABI for tools integration (#479)
One of the core components of Swift Testing is its ability to
interoperate with Xcode 16, VS Code, and other tools. Swift Testing has
been fully open-sourced across all platforms supported by Swift, and can
be added as a package dependency (or—eventually—linked from the Swift
toolchain.)

Because Swift Testing may be used in various forms, and because
integration with various tools is critical to its success, we need it to
have a stable interface that can be used regardless of how it's been
added to a package.

Read the full proposal
[here](https://github.com/apple/swift-testing/blob/main/Documentation/Proposals/0002-json-abi.md).

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.