Emit ABIv0 JSON objects for tests when listing them. (#511)
This PR changes the behavior of `swift test list` and its various
synonyms to allow reporting the list of tests via the ABI-stable JSON
mechanism described in #479.

As it is not currently possible to directly call `swift test list
--event-stream-output-path ... --event-stream-version 0`, it's a bit
hard to test this code. However, it is possible to opt into this mode
using `--configuration-path` and passing a path to a JSON file that
includes `"listTests": true` (as noted by @allevato.) Support for these
arguments with `swift test list` is tracked by
https://github.com/swiftlang/swift-package-manager/pull/7768.

Resolves #506.
Resolves rdar://130627856.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.