Do not write to `stderr` when testing `swift test list`. (#533)
When testing the JSON ABI _vis-à-vis_ `swift test list`, we're writing
to `stderr`. For the purposes of our tests, this is just noise, so block
output to `stderr` from the `list` subcommand if verbosity is set to
`.min` (as it is in the tests.)

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.