Lower `Backtrace` and `SourceContext` to SPI.
`Backtrace` is our implementation of a stack trace type. Xcode 16 uses it to
provide richer information about test failures than can be provided by _just_ an
instance of `SourceLocation`.

The standard library has its own type, `_Backtracing.Backtrace`, but due to our
deployment requirements on Apple platforms, we can't use it yet. It would make
sense for us to adopt the `Backtrace` type in the standard library as soon as
possible. To avoid staging issues in the future, and because our type has not
undergone any sort of formal review, I'm lowering it to tools-only SPI.

I'm also lowering `SourceContext` because it consists of an instance of
`Backtrace` and an instance of `SourceLocation`—but if `Backtrace` is SPI, then
the public interface of `SourceContext` is simply a wrapper around
`SourceLocation`. This does not eliminate the internal need for the type, but it
does make it redundant from an API perspective. If and when we regain
`Backtrace` as a public interface, we can look at raising `SourceContext` too.