Work around a bug in the older Swift compiler used to build swift-syntax in the toolchain, which affected certain result builder APIs (#541)
This works around a bug which was identified while working towards
adding swift-testing to Swift toolchains. See the code comment for a
more detailed explanation.

### Modifications:

Use a different variant of this swift-syntax API in our macro plugin to
avoid the bug.

### Result:

Building in the toolchain avoids the failure this was causing.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.