Allow non-`final` classes to act as suites. (#550)
Now that we require Swift 6, we can see the names of classes during
`@Test` macro expansion. This was the primary blocker preventing us from
allowing non-`final` classes, because we had to substitute `Self` for
the class name and this was invalid if the suite type could be
covariant. We can drop the constraint now.

Note that test functions are _not_ heritable.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.