Explicitly import Foundation in IssueTests.swift. (#567)
If XCTest is unavailable, IssueTests.swift fails to compile due to a
Foundation dependency (that was transitively satisfied by importing
XCTest which reexports Foundation). Add an explicit import of Foundation
in this file to resolve the issue.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.