Ensure serially-scheduled tests run in source order. (#569)
This PR ensures that tests, when scheduled serially, run in declared
source order (or as close to it as suite layout allows.)

When tests are run in parallel, we don't bother to sort them because
they will run non-deterministically anyway. Although we don't explicitly
shuffle them, they are stored in a `Dictionary` and in Swift, hashcodes
are salted with a random value, so their iteration order will change on
every test run. (In other words, they're already effectively shuffled.)

Resolves #566.
Resolves rdar://132285155.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.