Move the `stdout` line-buffering call. (#576)
Move the `stdout` line-buffering call so it affects both overloads of
`__swiftPMEntryPoint()`.

This has no effect on SwiftPM because it currently uses the overload
that returns `Never`. Two overloads are provided to allow SwiftPM some
flexibility, but we're not actively taking advantage of that in Swift 6.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.